Multipage functionality!
I added the crawl! It's probably not the fastest (opens and closes new instances of chrome) but it's the only thing I could do to bypass Cloudflare.

I need to add more pages in the URLs list but I'm not too sure what I want to scrape yet. This will be it until I work on the front-end, I have enough data to start working on that :)

TODO:
make a readme
clean up code (always)
test wait times
add more to URLs page